+++ hw06-solid/src/main/java/ru/otus/atm/processor/ProcessorImpl.java
@@ -1,4 +1,4 @@
-package ru.otus.atm.dispenser;
+package ru.otus.atm.processor;
 
 import java.util.ArrayList;
 import java.util.List;
