+++ hw06-solid/src/test/java/ru/otus/atm/noteholder/rf/RoubleBanknoteHolderTest.java
@@ -6,10 +6,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import ru.otus.atm.Atm;
 import ru.otus.atm.AtmService;
-import ru.otus.atm.dispenser.AmountConverter;
-import ru.otus.atm.dispenser.AmountConverterImpl;
-import ru.otus.atm.dispenser.Processor;
-import ru.otus.atm.dispenser.ProcessorImpl;
+import ru.otus.atm.processor.AmountConverter;
+import ru.otus.atm.processor.AmountConverterImpl;
+import ru.otus.atm.processor.Processor;
+import ru.otus.atm.processor.ProcessorImpl;
 import ru.otus.atm.noteholder.NoteHolder;
 import ru.otus.banknote.Banknote;
 import ru.otus.banknote.rf.Roubles500;
